pyproject.toml
[project]
name = "napari-slic"
version = "0.1.0"
description = "Interactive superpixel annotation tool using Napari and scikit-image SLIC."

requires-python = ">=3.10"

dependencies = [
    "napari>=0.5.0",
    "pyside6>=6.5",          # Qt backend with macOS/arm64 wheels (replaces napari[all])
    "scikit-image>=0.22",
    "numpy>=1.26",
    "magicgui>=0.8",
    "pyyaml>=6.0",           # config file loading
]

[project.scripts]
annotator = "annotator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
# Run directly without installing:  uv run annotator.py [image_path]
